reformatting and cleaning up the code
