+++ src/Services/jobseeker-service.js
@@ -1,63 +1,52 @@
 import { doc, setDoc, getDoc, getDocs, updateDoc, deleteDoc, collection } from "firebase/firestore";
-import firebase from '../../firebase'; //idk if i need these two lines yet
+import firebase from '../../firebase'; 
 var db = firebase
 
 // CRUD functions
-//data in form {"name": "John"} for instance.
+//data in form {"name": "John"} for instance; remember event.preventDefault() when using.
 export const createJobseeker = async (email, data) => {
-    await setDoc(doc(db, "jobseekers", email), data).then(function () {
-        console.log(`created new jobseeker`);
-    }).catch(function (err) {
-        alert(err.stack);
-    });
+  await setDoc(doc(db, "jobseekers", email), data).then(function () {
+      console.log(`created new jobseeker`);
+  }).catch(function (err) {
+      alert(err.stack);
+  });
 }
-
 
 export const fetchJobseeker = async (email) => {
-    const docRef = doc(db, "jobseekers", email);
-    try {
-        const docSnap = await getDoc(docRef);
-        if (docSnap.exists()) {
-        console.log("jobseeker: ", docSnap.data());
-        } else {
-        console.log("jobseeker ", email, " does not exist");
-        }
+  const docRef = doc(db, "jobseekers", email);
+  try {
+    const docSnap = await getDoc(docRef);
+    if (docSnap.exists()) {
+    console.log("jobseeker: ", docSnap.data());
+    } else {
+    console.log("jobseeker ", email, " does not exist");
     }
-    catch (error) {
-        console.log(error);
-    }
+  }
+  catch (error) {
+      console.log(error);
+  }
 }
 
- export const fetchAllJobseekers = async () => {
-    const colRef = collection(db, "jobseekers");
-    try {
-      const docsSnap = await getDocs(colRef);
-      return docsSnap;
-      // if(docsSnap.docs.length > 0) {
-      //    docsSnap.forEach(doc => {
-      //       console.log(doc.data());
-      //       console.log(doc.id);
-      //    })
-      // }
-    } catch (error) {
-        console.log(error);
-    }
+export const fetchAllJobseekers = async () => {
+  const colRef = collection(db, "jobseekers");
+  try {
+    const docsSnap = await getDocs(colRef);
+    return docsSnap;
+  } catch (error) {
+      console.log(error);
   }
+}
 
-  
-  export const updateJobseeker = async (email, data) => {
-    await updateDoc(doc(db, "jobseekers", email), data)
-    .then(function () {
-      console.log(`updated jobseeker `, email);
-    }).catch(function (err) {
-      alert(err.stack);
-    });
-  }
+export const updateJobseeker = async (email, data) => {
+  await updateDoc(doc(db, "jobseekers", email), data)
+  .then(function () {
+    console.log(`updated jobseeker `, email);
+  }).catch(function (err) {
+    alert(err.stack);
+  });
+}
 
 export const deleteJobseeker = async (email) => {
-    // make sure to use "read" to be sure the user exists in the database before calling delete
-    // this function alone won't be able to confirm if account has been deleted or doesnt exist
-    //i think its fine bc of try + catch.
     await deleteDoc(doc(db, "jobseekers", email)).then(() => {
         console.log("Jobseeker account ", email, " has been deleted successfully.")
     })
@@ -75,7 +64,7 @@ export const deleteJobseeker = async (email) => {
     milestoneArray[milestoneIndex] = "awaiting approval";
     await updateDoc(doc(db, "jobseekers", JSemail), {"milestoneArray": milestoneArray})
     .then(function () {
-      //email NavEmail here; may potentially have multiple navigators maybe?
+      //email NavEmail here; may potentially have multiple navigators(?)
       console.log(`pinged navigator for milestone approval `, JSemail);
     }).catch(function (err) {
       alert(err.stack);
